PR comment are resolved
