+++ src/app/authentication/business/changePasswordService.go
@@ -4,8 +4,8 @@ import (
 	"authentication/commons/constants"
 	"authentication/models"
 	"authentication/repositories"
-	"authentication/utils"
 	"errors"
+	"stock_broker_application/src/utils/validations"
 	"github.com/gin-gonic/gin"
 	"github.com/go-playground/validator/v10"
 )
@@ -30,7 +30,7 @@ func (service *PasswordResetter) ResetPassword(request models.ChangePassword, ct
 		return errors.New(constants.ErrorInvalidEmailOrPassword)
 	}
 	validate := validator.New()
-	if err := utils.RegisterCustomValidations(validate); err != nil {
+	if err := validations.RegisterCustomValidations(validate); err != nil {
 		return errors.New(err.Error())
 	}
 	if err := validate.Struct(request); err != nil {
