+++ src/app/authentication/commons/constants/constants.go
@@ -17,8 +17,3 @@ const (
 	EmailId  = "email"
 	Password = "password"
 )
-
-//password regex expression
-const (
-	PasswordRegex = `^[a-zA-Z0-9!@#$%^&*()_+=\-[\]{};:'",.<>/?]{8,}$`
-)
