+++ src/app/authentication/docs/swagger.json
@@ -65,13 +65,11 @@
             "properties": {
                 "newPassword": {
                     "type": "string",
-                    "minLength": 8,
-                    "example": "newPassword"
+                    "example": "Coditas@18"
                 },
                 "oldPassword": {
                     "type": "string",
-                    "minLength": 8,
-                    "example": "password"
+                    "example": "S@nket123"
                 }
             }
         }
